feat(ui): split homepage for TSM & IOx
